Install ginkgo, gomega, add line to readme, move tests to subpackage
