+++ README.md
@@ -25,3 +25,10 @@ Using the github cli:
 ```bash
 gh pr create
 ```
+
+## Running Tests
+
+Ensure you have the ginkgo testing framewokr installed:
+```bash
+go install github.com/onsi/ginkgo/ginkgo
+```
