+++ go.mod
@@ -1,3 +1,8 @@
 module github.com/StevensSEC/pokemonbattlelib
 
 go 1.15
+
+require (
+	github.com/onsi/ginkgo v1.15.0 // indirect
+	github.com/onsi/gomega v1.10.5 // indirect
+)
