+++ pokemonbattlelib_test/pokemonbattlelib_suite_test.go
@@ -1,0 +1,13 @@
+package pokemonbattlelib_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestPokemonbattlelib(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Pokemonbattlelib Suite")
+}
